Edits

+++ CodeBook.Business.App/Services/CommentService.cs
@@ -3,6 +3,7 @@
 using CodeBook.Business.App.DTOs;
 using CodeBook.Business.App.Interfaces;
 using CodeBook.Business.App.Interfaces;
+using CodeBook.Business.App.Middleware;
 using CodeBook.Data.App.IRepositories;
 using CodeBook.Data.App.IRepositories;
 using CodeBook.Data.App.Repositories;
@@ -19,22 +20,38 @@ namespace CodeBook.Business.App.Services
     {
         private readonly ICommentRepository _commentRepository;
         private readonly INotificationService _notificationService;
+        private readonly IPostRepository _postRepository;   
 
-        public CommentService(ICommentRepository commentRepository,INotificationService notificationService)
+        public CommentService(ICommentRepository commentRepository,INotificationService notificationService, IPostRepository postRepository)
         {
             _commentRepository = commentRepository;
             _notificationService = notificationService;
+            _postRepository = postRepository;
         }
 
-        public void AddComment(int authorId,int postId,AddCommentRequest dto)
+        public ErrorResponse AddComment(int authorId, int postId, AddCommentRequest dto)
         {
-            Comment comment = new Comment();
-            comment.AuthorId = authorId;
-            comment.PostId = postId;
-            comment.Body = dto.Body;
-            comment.SelfCommentId = dto.SelfCommentId;
-           _commentRepository.Add(comment);
-           _commentRepository.SaveChanges();
+            var post = _postRepository.GetPostById(postId);
+            if(post == null)
+            {
+                return new ErrorResponse { Success = false, Message = "Post not found"};
+            }
+            if (post.IsRemoved)
+            {
+                return new ErrorResponse { Success = false, Message = "Post is no longer available" };
+            }
+
+            Comment comment = new Comment
+            {
+                AuthorId = authorId,
+                PostId = postId,
+                Body = dto.Body,
+                SelfCommentId = dto.SelfCommentId,
+                DateCreated = DateTime.UtcNow,
+            };
+
+            if (!_commentRepository.SaveChanges())
+                return new ErrorResponse { Success = false, Message = "Could not add comment" };
 
             _notificationService.CreateNotification(authorId, new NotificationDTO
             {
@@ -45,26 +62,56 @@ namespace CodeBook.Business.App.Services
                 IsSeen = false,
                 DateCreated = DateTime.UtcNow
             });
+            return new ErrorResponse { Success = true, Message = "Comment added successfully" };
 
         }
-        public void EditComment(int commentId,string CommentBody)
+        public ErrorResponse EditComment(int commentId,string commentBody, int userId)
         {
-            Comment comment = _commentRepository.GetCommentById(commentId);
-            comment.Body = CommentBody;
+            var comment = _commentRepository.GetCommentById(commentId);
+            if (comment == null)
+            {
+                return new ErrorResponse { Success = false, Message = "Comment not found" };
+            }
+            if(comment.AuthorId != userId)
+            {
+                return new ErrorResponse { Success = false, Message = "You can only edit yoy own comments" };
+            }
+
+            comment.Body = commentBody;
+            
             _commentRepository.Update(comment);
-            _commentRepository.SaveChanges();
+            if (_commentRepository.SaveChanges())
+                return new ErrorResponse { Success = true, Message = "Comment updated successfully" };
+
+            return new ErrorResponse { Success = false, Message = "Could not update comment" };
 
         }
-        public void DeleteComment(int commentId)
+        public ErrorResponse DeleteComment(int commentId, int userId)
         {
-            Comment comment = _commentRepository.GetCommentById(commentId);
+            var comment = _commentRepository.GetCommentById(commentId);
+            if (comment == null)
+            {
+                return new ErrorResponse { Success = false, Message = "Comment not found" };
+            }
+            if (comment.AuthorId != userId)
+            {
+                return new ErrorResponse { Success = false, Message = "You can only delete your own comments" };
+            }
             _commentRepository.Delete(comment);
-            _commentRepository.SaveChanges();
+            if (_commentRepository.SaveChanges())
+                return new ErrorResponse { Success = true, Message = "Comment deleted successfully" };
+
+            return new ErrorResponse { Success = false, Message = "Could not delete comment" };
         }
 
         public List<CommentDto> GetPostComments(int postId)
         {
             var comments = _commentRepository.GetByPostId(postId);
+
+            if(comments == null || !comments.Any())
+            {
+                return new List<CommentDto>();
+            }
 
             return comments.Select(c => new CommentDto
             {
@@ -80,6 +127,10 @@ namespace CodeBook.Business.App.Services
         public int GetCommentAuthorId(int commentId)
         {
             var comment = _commentRepository.GetCommentById(commentId);
+            if (comment == null)
+            {
+                throw new Exception("Comment not found");
+            }
             return comment.AuthorId;
         }
     }
